pyproject.toml
[project]
name = "ride-share-simulator"
version = "0.1.0"
description = "match rides and drivers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.11.1",
    "fastapi>=0.116.1",
    "ijson>=3.4.0",
    "inotify>=0.2.12",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 122
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["ALL"]  # Enable all rules by default
ignore = [
    "D", # Ignore documentation warnings
    "ANN", # Ignore type annotation warnings temporarily
    "ERA", # Ignore commented-out code warnings
    "ARG", # Ignore unused arguments warnings
    "INP", # Ignore implicit namespace warnings
    "T20", # Ignore print statements
    "PLR0913", # Ignore too many arguments warning
    "C416" # Ignore list comprehension warnings
]

[tool.ruff.lint.pycodestyle]
max-line-length = 122

[tool.ruff.lint.mccabe]
max-complexity = 10

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]